fix: resolve all lint and type errors, add pre-commit hooks
- Fix 120 ruff lint errors (unused imports, line length, type annotations,
  raise-from, comprehensions, raw strings)
- Fix 61 mypy strict mode errors (type annotations, casts, generics)
- Add .pre-commit-config.yaml with trailing-whitespace, end-of-file-fixer,
  check-yaml/json/toml, merge-conflict, debug-statements, detect-private-key,
  ruff lint+format, and mypy strict
- Expand ruff rules: add C4, PIE, PT, RUF, PERF
- Add types-PyYAML and pre-commit to dev dependencies
- Configure mypy overrides for tomllib/sre_parse stub gaps

+++ crossfire/confidence.py
@@ -102,7 +102,7 @@ def recommend_samples(
         # All matches or no matches — minimal samples needed
         return max(10, total)
 
-    n = int(math.ceil(4 * (z / target_width) ** 2 * variance))
+    n = math.ceil(4 * (z / target_width) ** 2 * variance)
     return max(10, n)
